+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# 1.4.7
+Bug fix:
+- [#247](https://github.com/elysiajs/eden/pull/247), [#246](https://github.com/elysiajs/eden/issues/246) handle network error
+- [#244](https://github.com/elysiajs/eden/issues/244) don't skip index path name
+- [#241](https://github.com/elysiajs/eden/issues/241) prevent freeze from thenable
+- [#236](https://github.com/elysiajs/eden/issues/236) reconcile default headers in type-level
+- [#232](https://github.com/elysiajs/eden/issues/232) accept query array
+
 # 1.4.6 - 23 Dec 2025
 Bug fix:
 - [#235](https://github.com/elysiajs/eden/pull/235) SSE stream parser with partial chunks